crack_haus/requirements.txt
Django==1.7.6
django-filter==0.9.2
django-social-auth==0.7.28
djangorestframework==3.1.0
httplib2==0.9.1
Markdown==2.6.1
oauth2==1.5.211
Pillow==2.8.1
python-openid==2.2.5
